test/wpt/jsdom/issues/xxx.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>xxx</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<!-- Regression test for https://github.com/jsdom/jsdom/issues/xxx -->

<h1>Hello world</h1>
<h2>Goodbye, world</h2>

<script>
"use strict";
test(() => {
  assert_equals(document.querySelectorAll("h1,h2").length, 2);
});
</script>
